add showcase integration test
